fix(playerbot): Resolve covariant return types and duplicate destructors
- Moved AuctionMetrics, AuctionProfile, AuctionStrategy, TradeMetrics structs to interface
- Moved QuestPickupMetrics, ValidationMetrics structs to interface
- Renamed TradeSession in TradeManager.h to TradeManagerSession to avoid conflict
- Removed duplicate destructors from TradeSystem, QuestPickup, QuestValidation, RoleAssignment
- Added OPPORTUNISTIC strategy to AuctionStrategy enum
- All structs returned by value now have full definitions in interface headers
