+++ src/modules/Playerbot/Social/AuctionHouse.h
@@ -26,17 +26,7 @@
 namespace Playerbot
 {
 
-
-
-enum class AuctionActionType : uint8
-{
-    BUY_ITEM        = 0,
-    SELL_ITEM       = 1,
-    CANCEL_AUCTION  = 2,
-    UPDATE_BID      = 3,
-    SEARCH_MARKET   = 4,
-    ANALYZE_PRICES  = 5
-};
+// AuctionActionType enum defined in IAuctionHouse.h interface
 
 struct AuctionItem
 {
@@ -120,31 +110,7 @@ public:
     bool IsPriceBelowMarket(uint32 itemId, uint32 price, float threshold = 0.8f) override;
     void UpdateMarketData() override;
 
-    // Advanced auction features
-    struct AuctionProfile
-    {
-        AuctionStrategy primaryStrategy;
-        AuctionStrategy secondaryStrategy;
-        uint32 maxBiddingBudget;
-        uint32 maxListingBudget;
-        float bargainThreshold; // Buy if price is below this % of market value
-        float profitMargin; // Minimum profit margin for flipping
-        std::vector<uint32> preferredItemTypes;
-        std::vector<uint32> avoidedItemTypes;
-        std::unordered_set<uint32> watchList; // Items to monitor
-        std::unordered_set<uint32> blackList; // Never buy these items
-        bool autoRelist; // Automatically relist unsold items
-        bool autoBuyConsumables;
-        bool autoSellJunk;
-        uint32 maxAuctionsActive;
-
-        AuctionProfile() : primaryStrategy(AuctionStrategy::CONSERVATIVE)
-            , secondaryStrategy(AuctionStrategy::OPPORTUNISTIC), maxBiddingBudget(10000)
-            , maxListingBudget(5000), bargainThreshold(0.7f), profitMargin(0.2f)
-            , autoRelist(true), autoBuyConsumables(false), autoSellJunk(true)
-            , maxAuctionsActive(10) {}
-    };
-
+    // Advanced auction features (AuctionProfile defined in IAuctionHouse.h interface)
     void SetAuctionProfile(const AuctionProfile& profile);
     AuctionProfile GetAuctionProfile() const;
 
@@ -180,41 +146,7 @@ public:
     float GetCompetitorUndercutRate(uint32 sellerGuid) override;
     void TrackCompetitorBehavior(uint32 sellerGuid, const AuctionItem& auction) override;
 
-    // Performance monitoring
-    struct AuctionMetrics
-    {
-        std::atomic<uint32> auctionsCreated{0};
-        std::atomic<uint32> auctionsWon{0};
-        std::atomic<uint32> auctionsLost{0};
-        std::atomic<uint32> itemsSold{0};
-        std::atomic<uint32> itemsBought{0};
-        std::atomic<uint32> totalGoldSpent{0};
-        std::atomic<uint32> totalGoldEarned{0};
-        std::atomic<float> averageProfitMargin{0.2f};
-        std::atomic<float> successRate{0.8f};
-        std::atomic<uint32> marketScans{0};
-        std::chrono::steady_clock::time_point lastUpdate;
-
-        void Reset() {
-            auctionsCreated = 0; auctionsWon = 0; auctionsLost = 0;
-            itemsSold = 0; itemsBought = 0; totalGoldSpent = 0;
-            totalGoldEarned = 0; averageProfitMargin = 0.2f; successRate = 0.8f;
-            marketScans = 0; lastUpdate = std::chrono::steady_clock::now();
-        }
-
-        uint32 GetNetProfit() const {
-            uint32 earned = totalGoldEarned.load();
-            uint32 spent = totalGoldSpent.load();
-            return earned > spent ? earned - spent : 0;
-        }
-
-        float GetWinRate() const {
-            uint32 won = auctionsWon.load();
-            uint32 total = won + auctionsLost.load();
-            return total > 0 ? (float)won / total : 0.0f;
-        }
-    };
-
+    // Performance monitoring (AuctionMetrics defined in IAuctionHouse.h interface)
     AuctionMetrics GetAuctionMetrics() override;
     AuctionMetrics GetGlobalAuctionMetrics() override;
