+++ src/modules/Playerbot/Social/TradeSystem.h
@@ -193,34 +193,7 @@ public:
     void LogTradeTransaction(const TradeSession& session);
     void HandleTradeScamAttempt(Player* victim, Player* scammer);
 
-    // Performance monitoring
-    struct TradeMetrics
-    {
-        std::atomic<uint32> tradesInitiated{0};
-        std::atomic<uint32> tradesCompleted{0};
-        std::atomic<uint32> tradesCancelled{0};
-        std::atomic<uint32> vendorTransactions{0};
-        std::atomic<uint32> repairTransactions{0};
-        std::atomic<float> averageTradeValue{1000.0f};
-        std::atomic<float> tradeSuccessRate{0.8f};
-        std::atomic<uint32> totalGoldTraded{0};
-        std::atomic<uint32> totalItemsTraded{0};
-        std::chrono::steady_clock::time_point lastUpdate;
-
-        void Reset() {
-            tradesInitiated = 0; tradesCompleted = 0; tradesCancelled = 0;
-            vendorTransactions = 0; repairTransactions = 0; averageTradeValue = 1000.0f;
-            tradeSuccessRate = 0.8f; totalGoldTraded = 0; totalItemsTraded = 0;
-            lastUpdate = std::chrono::steady_clock::now();
-        }
-
-        float GetCompletionRate() const {
-            uint32 initiated = tradesInitiated.load();
-            uint32 completed = tradesCompleted.load();
-            return initiated > 0 ? (float)completed / initiated : 0.0f;
-        }
-    };
-
+    // Performance monitoring (TradeMetrics defined in ITradeSystem.h interface)
     TradeMetrics GetPlayerTradeMetrics() override;
     TradeMetrics GetGlobalTradeMetrics() override;
 
@@ -262,7 +235,6 @@ public:
 
 private:
     Player* _bot;
-    ~TradeSystem() = default;
 
     // Core data structures
     std::unordered_map<uint32, TradeSession> _activeTrades; // sessionId -> session
